set TERM to xterm
We want conservative terminal behavior anyway, but the old setting was also
causing the following error.

```
$ ./gradlew tasks

FAILURE: Build failed with an exception.

* What went wrong:
Could not open terminal for stdout: could not get termcap entry

* Try:
Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output.

* Get more help at https://help.gradle.org
```